+++ example/ios/Runner/AppDelegate.m
@@ -1,6 +1,11 @@
 #import "AppDelegate.h"
 #import "GeneratedPluginRegistrant.h"
-#import <SensorsAnalyticsSDK.h>
+#if __has_include(<SensorsAnalyticsSDK/SensorsAnalyticsSDK.h>)
+#import <SensorsAnalyticsSDK/SensorsAnalyticsSDK.h>
+#else
+#import "SensorsAnalyticsSDK.h"
+#endif
+
 #import <SensorsABTest.h>
 
 @implementation AppDelegate
@@ -9,14 +14,19 @@
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
   [GeneratedPluginRegistrant registerWithRegistry:self];
   // Override point for customization after application launch.
-    
-    // 测试环境，获取试验地址
-    NSString* kSABResultsTestURL = @"http://abtesting.saas.debugbox.sensorsdata.cn/api/v2/abtest/online/results?project-key=438B9364C98D54371751BA82F6484A1A03A5155E";
 
+//    [self startSensorsAnalyticsSDKWithOptions: launchOptions];
+//    
+//    [self startSensorsABTest];
+
+//
+    
+  return [super application:application didFinishLaunchingWithOptions:launchOptions];
+}
+
+- (void)startSensorsAnalyticsSDKWithOptions:(NSDictionary *)launchOptions {
     // 测试环境，数据接收地址
     NSString* kSABTestServerURL = @"http://10.130.6.4:8106/sa?project=default";
-
-    
     
     SAConfigOptions *options = [[SAConfigOptions alloc] initWithServerURL:kSABTestServerURL launchOptions:launchOptions];
 //    options.autoTrackEventType = SensorsAnalyticsEventTypeAppStart | SensorsAnalyticsEventTypeAppEnd | SensorsAnalyticsEventTypeAppClick | SensorsAnalyticsEventTypeAppViewScreen;
@@ -27,16 +37,23 @@
     options.enableVisualizedAutoTrack = YES;
     options.enableJavaScriptBridge = YES;
     options.enableLog = YES;
+    options.flushNetworkPolicy = SensorsAnalyticsNetworkTypeALL;
     [SensorsAnalyticsSDK startWithConfigOptions:options];
+    
+    
+    [SensorsAnalyticsSDK.sharedInstance track:@"test_track" withProperties:@{@"device_type": @"iPhone"}];
     
 
 //    [[SensorsAnalyticsSDK sharedInstance] setFlushNetworkPolicy:SensorsAnalyticsNetworkTypeALL];
-//
-    //SensorsABTestConfigOptions *abtestConfigOptions = [[SensorsABTestConfigOptions alloc] initWithURL:kSABResultsTestURL];
-    //[SensorsABTest startWithConfigOptions:abtestConfigOptions];
-//
+}
+
+- (void)startSensorsABTest {
+    // 测试环境，获取试验地址
+    NSString* kSABResultsTestURL = @"http://abtesting.saas.debugbox.sensorsdata.cn/api/v2/abtest/online/results?project-key=438B9364C98D54371751BA82F6484A1A03A5155E";
     
-  return [super application:application didFinishLaunchingWithOptions:launchOptions];
+    SensorsABTestConfigOptions *abtestConfigOptions = [[SensorsABTestConfigOptions alloc] initWithURL:kSABResultsTestURL];
+    [SensorsABTest startWithConfigOptions:abtestConfigOptions];
+    
 }
 
 @end