+++ CHANGELOG.md
@@ -1,12 +1,20 @@
-## 0.0.1
+## 1.0.0
 
-* Flutter A/B Testing 插件，支持 Android 和 iOS 中使用
+* 单个接口和全局设置，支持自定义属性请求分流
+* 支持自定义主体请求分流
 
+## 0.0.4
+
+* 修复 Android MethodChannel 偶现回调多次异常
+
+## 0.0.3
+
+* SDK 依赖由 jcenter 调整为 mavenCentral
 
 ## 0.0.2
 
 * 新增初始化接口
 
-## 0.0.3
+## 0.0.1
 
-* SDK 依赖由 jcenter 调整为 mavenCentral
+* Flutter A/B Testing 插件，支持 Android 和 iOS 中使用